+++ src/func_ov006_021051dc.c
@@ -13,7 +13,7 @@ void func_ov006_021051dc(char *c)
     int x;
 
     if (*(u16 *)(c + 0x4ec4) != 0) {
-        (*(u16 *)(int)(((long long)(int)(c + 0x4ec4))))--;
+        (*(u16 *)(int)(c + 0x4ec4))--;
         if (*(s16 *)(c + 0x4ec4) < 0)
             *(u16 *)(c + 0x4ec4) = 0;
         return;
@@ -49,7 +49,7 @@ void func_ov006_021051dc(char *c)
             int cell = *(int *)(c + 0x4cbc) * (row + y) + (col + x);
 
             *(u8 *)(c + cell + 0x4efa) = 1;
-            (*(u8 *)(int)(((long long)(int)(c + cell + 0x4f1e)))) ^= 1;
+            (*(u8 *)(int)(c + cell + 0x4f1e)) ^= 1;
             *(u16 *)(c + cell * 2 + 0x4e78) = 8;
         }
     }
@@ -57,5 +57,5 @@ void func_ov006_021051dc(char *c)
     func_ov006_02104cfc(c, idx);
     *(u16 *)(c + idx * 2 + 0x4e78) = 0;
     *(u16 *)(c + 0x4ec4) = 0x30;
-    (*(u8 *)(int)(((long long)(int)(c + 0x4fe5))))++;
+    (*(u8 *)(int)(c + 0x4fe5))++;
 }
