+++ src/func_ov020_021115ac.c
@@ -3,7 +3,7 @@
  * reserve it with no emitted code (plain or non-address-taken volatiles are DCE'd before
  * frame layout). The 6g u64-mask launder on &vel forces the ROM's `add r0,r4,#0xa4` base. */
 typedef struct { int x, y, z; } Vec3;
-#define LDR(p) ((int)((((long long)(int)(p)))))
+#define LDR(p) ((int)(p))
 
 extern void *_ZN5Actor10FindWithIDEj(u32 id);
 extern int _ZN5Actor24BumpedUnderneathByPlayerER6Player(void *thiz, void *player);
