+++ src/func_ov030_02112ff8.c
@@ -4,7 +4,7 @@ extern void _ZN12CylinderClsn5ClearEv(void *p);
 
 int func_ov030_02112ff8(char *c)
 {
-    *(int *)(((long long)(int)(c + 0xb0))) &= ~0x80000;
+    *(int *)(c + 0xb0) &= ~0x80000;
     if (Vec3_Dist(c + 0x380, c + 0x5c) < 0x514000 &&
         *(int *)(c + 0x60) > *(int *)(c + 0x384) - 0x12c000) {
         *(unsigned char *)(c + 0x3c7) = 0;
