+++ src/func_ov060_021132a4.c
@@ -3,7 +3,7 @@ typedef struct { int x, y, z; } Vector3;
 extern int _ZN8Particle6System3NewEjj5Fix12IiES2_S2_PK11Vector3_16fPNS_8CallbackE(unsigned int a, unsigned int b, int fix, int t1, int t2, void *v, void *cb);
 extern int _ZN5Sound8PlayLongEjjjRK7Vector3s(unsigned int a, unsigned int b, unsigned int c, void *v, unsigned int d);
 
-#define M(p) ((long long)(int)(p))
+#define M(p) (p)
 
 int func_ov060_021132a4(char *c)
 {
